+++ app/build.gradle
@@ -112,7 +112,8 @@ dependencies {
     implementation 'com.google.android.exoplayer:extension-okhttp:2.9.1'
 
     //gson
-    implementation 'com.squareup.retrofit2:converter-gson:2.5.0'
+    implementation 'com.github.campusappcn.AndRouter:router:1.2.8'
+    annotationProcessor 'com.github.campusappcn.AndRouter:compiler:1.2.8'
     ankoLibs.each { name ->
         implementation "org.jetbrains.anko:anko-$name:$ankoVersion"
     }
